Wrapper function for preparing R package with KWB styling

Wrapper function for preparing R package with KWB styling

use_pkg(author = list(name = "Michael Rustler", orcid = "0000-0003-0647-7726",
  url = "http://mrustl.de"), pkg = list(name = "kwb.lca", title =
  "R package supporting UMERTO LCA at KWB", desc =
  paste0("Helper functions for data aggregation and visualisation",
  " of UMBERTO (https://www.ifu.com/umberto/) model output.")),
  version = NULL, license = "MIT + file LICENSE",
  copyright_holder = list(name = "Kompetenzzentrum Wasser Berlin gGmbH (KWB)",
  start_year = NULL), funder = NULL, user = "KWB-R", domain = "github",
  stage = "experimental")

Arguments

author
author information in list format (default: list(name = "Michael Rustler", orcid = "0000-0003-0647-7726", url = "http://mrustl.de"))

pkg
package description in list format (default:
pkg = list(name = "kwb.lca",
title = "R package supporting UMERTO LCA at KWB",
desc = "Helper functions for data aggregation and visualisation of
UMBERTO (https://www.ifu.com/umberto) model output"))

version
user defined version number (e.g. 0.1.0) or default version number 0.0.0.9000 in case (version = NULL) of first release

license
license (default: "MIT + file LICENSE")

copyright_holder
list with name of copyright holder and additional start_year (e.g. "2017") of copyright protection. If not specified (start_year = NULL), the current year is used as year only! (default: "list(name ="Kompetenzzentrum Wasser Berlin gGmbH (KWB)", start_year = NULL)")

funder
funder/funding agengy of R package (default: NULL), e.g. project name (e.g. AQUANES)

user
user name or organisation under which repository defined in parameter "repo" is hosted (default: "KWB-R")

domain
under which repository is hosted (default: "github")

stage
badge declares the developmental stage of a package, according to [https://www.tidyverse.org/lifecycle/](https://www.tidyverse.org/lifecycle/), valid arguments are: "experimental", "maturing", "stable", "retired", "archived", "dormant", "questioning"), (default: "experiment")

Value

writes DESCRIPTION file using usethis::use_description() with KWB style